[c10] Move P2P access logic from ATen to c10
Refactor PeerToPeerAccess by moving the core implementation from
aten/src/ATen/cuda to c10/cuda. This makes P2P and fabric access
queries available at the c10 layer without requiring ATen dependencies.
The ATen layer now provides thin wrappers that ensure CUDA lazy
initialization before forwarding to c10. This separation allows
lower-level CUDA code to query P2P capabilities without pulling in
ATen context machinery.

Differential Revision: [D92675476](https://our.internmc.facebook.com/intern/diff/D92675476/)

[ghstack-poisoned]
