+++ c10/cuda/CMakeLists.txt
@@ -34,6 +34,7 @@ set(C10_CUDA_SRCS
     CUDAMallocAsyncAllocator.cpp
     CUDAMiscFunctions.cpp
     CUDAStream.cpp
+    PeerToPeerAccess.cpp
     impl/CUDAGuardImpl.cpp
     impl/CUDATest.cpp
     driver_api.cpp
@@ -50,6 +51,7 @@ set(C10_CUDA_HEADERS
     CUDAMathCompat.h
     CUDAMiscFunctions.h
     CUDAStream.h
+    PeerToPeerAccess.h
     impl/CUDAGuardImpl.h
     impl/CUDATest.h
 )
